refactor(cli): extract createAgent factory to eliminate duplicated agent construction
